Redirect anon. users to login when they don't have permission
For permissions that allow anonymous users, if an anonymous user
doesn't have permission to access a resource, redirect the user to the
login page.

This is useful if a user logs out and later returns to a view that
allows anonymous as well as authenticated users. For example, anons
might have access to published articles but not unpublished ones.