Add views to PermissionsRegistry Entry records
This is a set that keeps track of which views require a given
permission. It can be used to test/verify that a view does in fact
require a permission, which is the intended use case for this.

The PermissionsRegistry.entry_for_view() method was also added to
simplify fetching the Entry for a view/perm.