Make filters return False when first arg isn't a user
The first arg to template filters must now be an instance of either
get_user_model() or AnonymousUser (from django.contrib.auth).

The intent is to make templates more robust by not failing if a non-user
is passed as the first arg.

This might also be useful when passing a related user that might be
nullable (e.g., `some_object.user|can_do_something`).