Return wrapped permission function when registering
Previously, the original permission function would be returned when
registered via `PermissionsRegistry.register()`. The problem with that
was that if the permission function was called directly (e.g., in view
code), the `allow_anonymous`, `allow_staff`, and `allow_superuser`
options would be ignored.

We now return the wrapped permission function that checks these options
(the same wrapped function that's used for the template filter). So now
calling a permission function as a view decorator, template filter, or
directly will work the same way.