Make wrapped perm functions return False only when first arg is None
Previously, these would return False when the first arg was either None
or any other unexpected type. The intended purpose was to make templates
more robust, but I think that's a bad idea. We shouldn't be trying to
cover programming errors, and we shouldn't disallow duck-typing when it
makes sense (given that this is Python and all).

Amends dadbf2505df135870f8aa18f3533a8b37bee3582